+++ CHANGELOG.md
@@ -4,10 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.3.0 - 2021-04-22
+## 1.4.0 2021-04-28
 ### Added
 - new pro `manyToMany` for relations.
 - new way of relations many-to-many
+- Added middlewares to method `createStoreModule` inside `VuexOffline`
+  - fetchListQuery;
+  - fetchListSuccess;
+  - fetchListError;
+  - fetchFiltersSuccess;
+  - fetchFiltersError;
+  - fetchSingleSuccess;
+  - fetchSingleFormSuccess;
+  - fetchSingleError;
+  - saveSuccess;
+  - saveError;
+  - createSuccess;
+  - createError.
 
 ## 1.2.1 - 2021-04-16
 ### Added